Function derive_common::cg::propagate_clauses_to_output_type

pub fn propagate_clauses_to_output_type(
    where_clause: &mut Option<WhereClause>, 
    generics: &Generics, 
    trait_path: &Path, 
    trait_output: &Ident
)

Given an input type which has some where clauses already, like:

struct InputType where T: Zero, { ... }

Add the necessary where clauses so that the output type of a trait fulfils them.

For example:

    <T as ToComputedValue>::ComputedValue: Zero,

This needs to run before adding other bounds to the type parameters.